More VM stuff; basic page fault handler
- Better exception handlers; for page faults, we print the faulting
  address and saved processor state. This needs to be expanded for other
  types of exceptions though

- Remap newlines to CR+LF in PC UART output

- Recursive mapping for updating page tables (when the mapping is
  currently activated)

- Remap physical usage bitmaps as needed

- Added documentation for x86 arch memory map

+++ kernel/arch/x86/src/idt.c
@@ -32,7 +32,7 @@ void idt_init() {
 /**
  * Sets the value of an IDT gate.
  */
-void idt_set_entry(uint8_t entry, uint32_t function, uint8_t segment, uint8_t flags) {
+void idt_set_entry(uint8_t entry, uintptr_t function, uint8_t segment, uint8_t flags) {
     idt_entry_t *ptr = (idt_entry_t *) &sys_idt;
 
     ptr[entry].offset_1 = function & 0xFFFF;
